PR(-): Rename & reorganize doc actor relationship tests
