+++ tests/integration/acp/relationship/doc_actor/add/with_dummy_relation_test.go
@@ -8,7 +8,7 @@
 // by the Apache License, Version 2.0, included in the file
 // licenses/APL.txt.
 
-package test_acp_relationship_add_docactor
+package test_acp_relationship_doc_actor_add
 
 import (
 	"fmt"
@@ -19,12 +19,12 @@ import (
 	testUtils "github.com/sourcenetwork/defradb/tests/integration"
 )
 
-func TestACP_OwnerGivesDeleteWriteAccessToAnotherActorTwice_ShowThatTheRelationshipAlreadyExists(t *testing.T) {
+func TestACP_AddDocActorRelationshipWithDummyRelationDefinedOnPolicy_NothingChanges(t *testing.T) {
 	expectedPolicyID := "fc56b7509c20ac8ce682b3b9b4fdaad868a9c70dda6ec16720298be64f16e9a4"
 
 	test := testUtils.TestCase{
 
-		Description: "Test acp, owner gives write(delete) access to another actor twice, no-op",
+		Description: "Test acp, add doc actor relationship with a dummy relation defined on policy, nothing happens",
 
 		Actions: []any{
 			testUtils.AddPolicy{
@@ -119,18 +119,8 @@ func TestACP_OwnerGivesDeleteWriteAccessToAnotherActorTwice_ShowThatTheRelations
 				`,
 
 				Results: map[string]any{
-					"Users": []map[string]any{}, // Can't see the documents yet
+					"Users": []map[string]any{}, // Can't see the documents
 				},
-			},
-
-			testUtils.DeleteDoc{
-				CollectionID: 0,
-
-				Identity: immutable.Some(2), // This identity can not delete yet.
-
-				DocID: 0,
-
-				ExpectedError: "document not found or not authorized to access",
 			},
 
 			testUtils.AddDocActorRelationship{
@@ -142,23 +132,27 @@ func TestACP_OwnerGivesDeleteWriteAccessToAnotherActorTwice_ShowThatTheRelations
 
 				DocID: 0,
 
-				Relation: "writer",
+				Relation: "dummy", // Doesn't mean anything to the database.
 
 				ExpectedExistence: false,
 			},
 
-			testUtils.AddDocActorRelationship{
-				RequestorIdentity: 1,
+			testUtils.Request{
+				Identity: immutable.Some(2), // This identity can still not read.
 
-				TargetIdentity: 2,
+				Request: `
+					query {
+						Users {
+							_docID
+							name
+							age
+						}
+					}
+				`,
 
-				CollectionID: 0,
-
-				DocID: 0,
-
-				Relation: "writer",
-
-				ExpectedExistence: true, // is a no-op
+				Results: map[string]any{
+					"Users": []map[string]any{}, // Can't see the documents
+				},
 			},
 		},
 	}
@@ -166,12 +160,12 @@ func TestACP_OwnerGivesDeleteWriteAccessToAnotherActorTwice_ShowThatTheRelations
 	testUtils.ExecuteTestCase(t, test)
 }
 
-func TestACP_OwnerGivesDeleteWriteAccessToAnotherActor_OtherActorCanDelete(t *testing.T) {
+func TestACP_AddDocActorRelationshipWithDummyRelationNotDefinedOnPolicy_Error(t *testing.T) {
 	expectedPolicyID := "fc56b7509c20ac8ce682b3b9b4fdaad868a9c70dda6ec16720298be64f16e9a4"
 
 	test := testUtils.TestCase{
 
-		Description: "Test acp, owner gives write(delete) access to another actor",
+		Description: "Test acp, add doc actor relationship with an invalid relation (not defined on policy), error",
 
 		Actions: []any{
 			testUtils.AddPolicy{
@@ -266,18 +260,8 @@ func TestACP_OwnerGivesDeleteWriteAccessToAnotherActor_OtherActorCanDelete(t *te
 				`,
 
 				Results: map[string]any{
-					"Users": []map[string]any{}, // Can't see the documents yet
+					"Users": []map[string]any{}, // Can't see the documents
 				},
-			},
-
-			testUtils.DeleteDoc{
-				CollectionID: 0,
-
-				Identity: immutable.Some(2), // This identity can not delete yet.
-
-				DocID: 0,
-
-				ExpectedError: "document not found or not authorized to access",
 			},
 
 			testUtils.AddDocActorRelationship{
@@ -289,13 +273,13 @@ func TestACP_OwnerGivesDeleteWriteAccessToAnotherActor_OtherActorCanDelete(t *te
 
 				DocID: 0,
 
-				Relation: "writer",
+				Relation: "NotOnPolicy", // Doesn't mean anything to the database and not on policy either.
 
-				ExpectedExistence: false,
+				ExpectedError: "failed to add document actor relationship with acp",
 			},
 
 			testUtils.Request{
-				Identity: immutable.Some(2), // This identity can now read.
+				Identity: immutable.Some(2), // This identity can still not read.
 
 				Request: `
 					query {
@@ -308,194 +292,7 @@ func TestACP_OwnerGivesDeleteWriteAccessToAnotherActor_OtherActorCanDelete(t *te
 				`,
 
 				Results: map[string]any{
-					"Users": []map[string]any{
-						{
-							"_docID": "bae-9d443d0c-52f6-568b-8f74-e8ff0825697b",
-							"name":   "Shahzad",
-							"age":    int64(28),
-						},
-					},
-				},
-			},
-
-			testUtils.DeleteDoc{
-				CollectionID: 0,
-
-				Identity: immutable.Some(2), // This identity can now delete.
-
-				DocID: 0,
-			},
-
-			testUtils.Request{
-				Identity: immutable.Some(2), // Check if actually deleted.
-
-				Request: `
-					query {
-						Users {
-							_docID
-							name
-							age
-						}
-					}
-				`,
-
-				Results: map[string]any{
-					"Users": []map[string]any{},
-				},
-			},
-		},
-	}
-
-	testUtils.ExecuteTestCase(t, test)
-}
-
-func TestACP_OwnerGivesDeleteWriteAccessToAnotherActor_OtherActorCanDeleteSoCanTheOwner(t *testing.T) {
-	expectedPolicyID := "fc56b7509c20ac8ce682b3b9b4fdaad868a9c70dda6ec16720298be64f16e9a4"
-
-	test := testUtils.TestCase{
-
-		Description: "Test acp, owner gives write(delete) access to another actor, both can read",
-
-		Actions: []any{
-			testUtils.AddPolicy{
-
-				Identity: immutable.Some(1),
-
-				Policy: `
-                    name: Test Policy
-
-                    description: A Policy
-
-                    actor:
-                      name: actor
-
-                    resources:
-                      users:
-                        permissions:
-                          read:
-                            expr: owner + reader + writer
-
-                          write:
-                            expr: owner + writer
-
-                          nothing:
-                            expr: dummy
-
-                        relations:
-                          owner:
-                            types:
-                              - actor
-
-                          reader:
-                            types:
-                              - actor
-
-                          writer:
-                            types:
-                              - actor
-
-                          admin:
-                            manages:
-                              - reader
-                            types:
-                              - actor
-
-                          dummy:
-                            types:
-                              - actor
-                `,
-
-				ExpectedPolicyID: expectedPolicyID,
-			},
-
-			testUtils.SchemaUpdate{
-				Schema: fmt.Sprintf(`
-						type Users @policy(
-							id: "%s",
-							resource: "users"
-						) {
-							name: String
-							age: Int
-						}
-					`,
-					expectedPolicyID,
-				),
-			},
-
-			testUtils.CreateDoc{
-				Identity: immutable.Some(1),
-
-				CollectionID: 0,
-
-				Doc: `
-					{
-						"name": "Shahzad",
-						"age": 28
-					}
-				`,
-			},
-
-			testUtils.AddDocActorRelationship{
-				RequestorIdentity: 1,
-
-				TargetIdentity: 2,
-
-				CollectionID: 0,
-
-				DocID: 0,
-
-				Relation: "writer",
-
-				ExpectedExistence: false,
-			},
-
-			testUtils.Request{
-				Identity: immutable.Some(1), // Owner can still also delete (ownership not transferred)
-
-				Request: `
-					query {
-						Users {
-							_docID
-							name
-							age
-						}
-					}
-				`,
-
-				Results: map[string]any{
-					"Users": []map[string]any{
-						{
-							"_docID": "bae-9d443d0c-52f6-568b-8f74-e8ff0825697b",
-							"name":   "Shahzad",
-							"age":    int64(28),
-						},
-					},
-				},
-			},
-
-			testUtils.DeleteDoc{
-				CollectionID: 0,
-
-				Identity: immutable.Some(1), // Owner can still also delete.
-
-				DocID: 0,
-			},
-
-			testUtils.Request{
-				Identity: immutable.Some(1), // Check if actually deleted.
-
-				Request: `
-					query {
-						Users {
-							_docID
-							name
-							age
-						}
-					}
-				`,
-
-				Results: map[string]any{
-					"Users": []map[string]any{},
+					"Users": []map[string]any{}, // Can't see the documents
 				},
 			},
 		},
