+++ tests/integration/acp/relationship/doc_actor/add/with_reader_test.go
@@ -8,7 +8,7 @@
 // by the Apache License, Version 2.0, included in the file
 // licenses/APL.txt.
 
-package test_acp_relationship_add_docactor
+package test_acp_relationship_doc_actor_add
 
 import (
 	"fmt"
@@ -19,7 +19,144 @@ import (
 	testUtils "github.com/sourcenetwork/defradb/tests/integration"
 )
 
-func TestACP_ManagerGivesReadAccessToAnotherActor_OtherActorCanRead(t *testing.T) {
+func TestACP_OwnerGivesReadAccessToAnotherActorTwice_ShowThatTheRelationshipAlreadyExists(t *testing.T) {
+	expectedPolicyID := "fc56b7509c20ac8ce682b3b9b4fdaad868a9c70dda6ec16720298be64f16e9a4"
+
+	test := testUtils.TestCase{
+
+		Description: "Test acp, owner gives read access to another actor twice, no-op",
+
+		Actions: []any{
+			testUtils.AddPolicy{
+
+				Identity: immutable.Some(1),
+
+				Policy: `
+                    name: Test Policy
+
+                    description: A Policy
+
+                    actor:
+                      name: actor
+
+                    resources:
+                      users:
+                        permissions:
+                          read:
+                            expr: owner + reader + writer
+
+                          write:
+                            expr: owner + writer
+
+                          nothing:
+                            expr: dummy
+
+                        relations:
+                          owner:
+                            types:
+                              - actor
+
+                          reader:
+                            types:
+                              - actor
+
+                          writer:
+                            types:
+                              - actor
+
+                          admin:
+                            manages:
+                              - reader
+                            types:
+                              - actor
+
+                          dummy:
+                            types:
+                              - actor
+                `,
+
+				ExpectedPolicyID: expectedPolicyID,
+			},
+
+			testUtils.SchemaUpdate{
+				Schema: fmt.Sprintf(`
+						type Users @policy(
+							id: "%s",
+							resource: "users"
+						) {
+							name: String
+							age: Int
+						}
+					`,
+					expectedPolicyID,
+				),
+			},
+
+			testUtils.CreateDoc{
+				Identity: immutable.Some(1),
+
+				CollectionID: 0,
+
+				Doc: `
+					{
+						"name": "Shahzad",
+						"age": 28
+					}
+				`,
+			},
+
+			testUtils.Request{
+				Identity: immutable.Some(2), // This identity can not read yet.
+
+				Request: `
+					query {
+						Users {
+							_docID
+							name
+							age
+						}
+					}
+				`,
+
+				Results: map[string]any{
+					"Users": []map[string]any{}, // Can't see the documents yet
+				},
+			},
+
+			testUtils.AddDocActorRelationship{
+				RequestorIdentity: 1,
+
+				TargetIdentity: 2,
+
+				CollectionID: 0,
+
+				DocID: 0,
+
+				Relation: "reader",
+
+				ExpectedExistence: false,
+			},
+
+			testUtils.AddDocActorRelationship{
+				RequestorIdentity: 1,
+
+				TargetIdentity: 2,
+
+				CollectionID: 0,
+
+				DocID: 0,
+
+				Relation: "reader",
+
+				ExpectedExistence: true, // is a no-op
+			},
+		},
+	}
+
+	testUtils.ExecuteTestCase(t, test)
+}
+
+func TestACP_OwnerGivesReadAccessToAnotherActor_OtherActorCanRead(t *testing.T) {
 	expectedPolicyID := "fc56b7509c20ac8ce682b3b9b4fdaad868a9c70dda6ec16720298be64f16e9a4"
 
 	test := testUtils.TestCase{
@@ -106,7 +243,7 @@ func TestACP_ManagerGivesReadAccessToAnotherActor_OtherActorCanRead(t *testing.T
 			},
 
 			testUtils.Request{
-				Identity: immutable.Some(3), // This identity can not read yet.
+				Identity: immutable.Some(2), // This identity can not read yet.
 
 				Request: `
 					query {
@@ -123,24 +260,10 @@ func TestACP_ManagerGivesReadAccessToAnotherActor_OtherActorCanRead(t *testing.T
 				},
 			},
 
-			testUtils.AddDocActorRelationship{ // Make admin / manager
+			testUtils.AddDocActorRelationship{
 				RequestorIdentity: 1,
 
 				TargetIdentity: 2,
-
-				CollectionID: 0,
-
-				DocID: 0,
-
-				Relation: "admin",
-
-				ExpectedExistence: false,
-			},
-
-			testUtils.AddDocActorRelationship{ // Admin makes another actor a reader
-				RequestorIdentity: 2,
-
-				TargetIdentity: 3,
 
 				CollectionID: 0,
 
@@ -152,7 +275,7 @@ func TestACP_ManagerGivesReadAccessToAnotherActor_OtherActorCanRead(t *testing.T
 			},
 
 			testUtils.Request{
-				Identity: immutable.Some(3), // The other actor can read
+				Identity: immutable.Some(2), // Now this identity can read.
 
 				Request: `
 					query {
@@ -174,1016 +297,20 @@ func TestACP_ManagerGivesReadAccessToAnotherActor_OtherActorCanRead(t *testing.T
 					},
 				},
 			},
-
-			testUtils.UpdateDoc{
-				CollectionID: 0,
-
-				Identity: immutable.Some(3), // The other actor can not update
-
-				DocID: 0,
-
-				Doc: `
-					{
-						"name": "Shahzad Lone"
-					}
-				`,
-
-				ExpectedError: "document not found or not authorized to access",
-			},
-
-			testUtils.DeleteDoc{
-				CollectionID: 0,
-
-				Identity: immutable.Some(3), // The other actor can not delete
-
-				DocID: 0,
-
-				ExpectedError: "document not found or not authorized to access",
-			},
 		},
 	}
 
 	testUtils.ExecuteTestCase(t, test)
 }
 
-func TestACP_ManagerGivesWriteAccessToAnotherActor_OtherActorCanWrite(t *testing.T) {
+// Note: Testing that owner can still read after the relationship was formed is to ensure
+// that no transfer of ownership has taken place.
+func TestACP_OwnerGivesReadAccessToAnotherActor_OtherActorCanReadSoCanTheOwner(t *testing.T) {
 	expectedPolicyID := "fc56b7509c20ac8ce682b3b9b4fdaad868a9c70dda6ec16720298be64f16e9a4"
 
 	test := testUtils.TestCase{
 
-		Description: "Test acp, owner gives write access to another actor",
-
-		Actions: []any{
-			testUtils.AddPolicy{
-
-				Identity: immutable.Some(1),
-
-				Policy: `
-                    name: Test Policy
-
-                    description: A Policy
-
-                    actor:
-                      name: actor
-
-                    resources:
-                      users:
-                        permissions:
-                          read:
-                            expr: owner + reader + writer
-
-                          write:
-                            expr: owner + writer
-
-                          nothing:
-                            expr: dummy
-
-                        relations:
-                          owner:
-                            types:
-                              - actor
-
-                          reader:
-                            types:
-                              - actor
-
-                          writer:
-                            types:
-                              - actor
-
-                          admin:
-                            manages:
-                              - writer
-                            types:
-                              - actor
-
-                          dummy:
-                            types:
-                              - actor
-                `,
-
-				ExpectedPolicyID: expectedPolicyID,
-			},
-
-			testUtils.SchemaUpdate{
-				Schema: fmt.Sprintf(`
-						type Users @policy(
-							id: "%s",
-							resource: "users"
-						) {
-							name: String
-							age: Int
-						}
-					`,
-					expectedPolicyID,
-				),
-			},
-
-			testUtils.CreateDoc{
-				Identity: immutable.Some(1),
-
-				CollectionID: 0,
-
-				Doc: `
-					{
-						"name": "Shahzad",
-						"age": 28
-					}
-				`,
-			},
-
-			testUtils.Request{
-				Identity: immutable.Some(3), // This identity can not read yet.
-
-				Request: `
-					query {
-						Users {
-							_docID
-							name
-							age
-						}
-					}
-				`,
-
-				Results: map[string]any{
-					"Users": []map[string]any{}, // Can't see the documents yet
-				},
-			},
-
-			testUtils.AddDocActorRelationship{ // Make admin / manager
-				RequestorIdentity: 1,
-
-				TargetIdentity: 2,
-
-				CollectionID: 0,
-
-				DocID: 0,
-
-				Relation: "admin",
-
-				ExpectedExistence: false,
-			},
-
-			testUtils.AddDocActorRelationship{ // Admin makes another actor a writer
-				RequestorIdentity: 2,
-
-				TargetIdentity: 3,
-
-				CollectionID: 0,
-
-				DocID: 0,
-
-				Relation: "writer",
-
-				ExpectedExistence: false,
-			},
-
-			testUtils.UpdateDoc{
-				CollectionID: 0,
-
-				Identity: immutable.Some(3), // The other actor can update
-
-				DocID: 0,
-
-				Doc: `
-					{
-						"name": "Shahzad Lone"
-					}
-				`,
-			},
-
-			testUtils.Request{
-				Identity: immutable.Some(3), // The other actor can read
-
-				Request: `
-					query {
-						Users {
-							_docID
-							name
-							age
-						}
-					}
-				`,
-
-				Results: map[string]any{
-					"Users": []map[string]any{
-						{
-							"_docID": "bae-9d443d0c-52f6-568b-8f74-e8ff0825697b",
-							"name":   "Shahzad Lone", // Updated name
-							"age":    int64(28),
-						},
-					},
-				},
-			},
-
-			testUtils.DeleteDoc{
-				CollectionID: 0,
-
-				Identity: immutable.Some(3), // The other actor can delete
-
-				DocID: 0,
-			},
-
-			testUtils.Request{
-				Identity: immutable.Some(3),
-
-				Request: `
-					query {
-						Users {
-							_docID
-							name
-							age
-						}
-					}
-				`,
-
-				Results: map[string]any{ // Check actually deleted
-					"Users": []map[string]any{},
-				},
-			},
-		},
-	}
-
-	testUtils.ExecuteTestCase(t, test)
-}
-
-func TestACP_OwnerMakesAManagerThatGivesItSelfReadAccess_ManagerCanRead(t *testing.T) {
-	expectedPolicyID := "fc56b7509c20ac8ce682b3b9b4fdaad868a9c70dda6ec16720298be64f16e9a4"
-
-	test := testUtils.TestCase{
-
-		Description: "Test acp, owner makes a manager that gives itself read access",
-
-		Actions: []any{
-			testUtils.AddPolicy{
-
-				Identity: immutable.Some(1),
-
-				Policy: `
-                    name: Test Policy
-
-                    description: A Policy
-
-                    actor:
-                      name: actor
-
-                    resources:
-                      users:
-                        permissions:
-                          read:
-                            expr: owner + reader + writer
-
-                          write:
-                            expr: owner + writer
-
-                          nothing:
-                            expr: dummy
-
-                        relations:
-                          owner:
-                            types:
-                              - actor
-
-                          reader:
-                            types:
-                              - actor
-
-                          writer:
-                            types:
-                              - actor
-
-                          admin:
-                            manages:
-                              - reader
-                            types:
-                              - actor
-
-                          dummy:
-                            types:
-                              - actor
-                `,
-
-				ExpectedPolicyID: expectedPolicyID,
-			},
-
-			testUtils.SchemaUpdate{
-				Schema: fmt.Sprintf(`
-						type Users @policy(
-							id: "%s",
-							resource: "users"
-						) {
-							name: String
-							age: Int
-						}
-					`,
-					expectedPolicyID,
-				),
-			},
-
-			testUtils.CreateDoc{
-				Identity: immutable.Some(1),
-
-				CollectionID: 0,
-
-				Doc: `
-					{
-						"name": "Shahzad",
-						"age": 28
-					}
-				`,
-			},
-
-			testUtils.Request{
-				Identity: immutable.Some(2), // This identity (to be manager) can not read yet.
-
-				Request: `
-					query {
-						Users {
-							_docID
-							name
-							age
-						}
-					}
-				`,
-
-				Results: map[string]any{
-					"Users": []map[string]any{}, // Can't see the documents yet
-				},
-			},
-
-			testUtils.AddDocActorRelationship{ // Make admin / manager
-				RequestorIdentity: 1,
-
-				TargetIdentity: 2,
-
-				CollectionID: 0,
-
-				DocID: 0,
-
-				Relation: "admin",
-
-				ExpectedExistence: false,
-			},
-
-			testUtils.AddDocActorRelationship{ // Manager makes itself a reader
-				RequestorIdentity: 2,
-
-				TargetIdentity: 2,
-
-				CollectionID: 0,
-
-				DocID: 0,
-
-				Relation: "reader",
-
-				ExpectedExistence: false,
-			},
-
-			testUtils.Request{
-				Identity: immutable.Some(2), // Manager can read now
-
-				Request: `
-					query {
-						Users {
-							_docID
-							name
-							age
-						}
-					}
-				`,
-
-				Results: map[string]any{
-					"Users": []map[string]any{
-						{
-							"_docID": "bae-9d443d0c-52f6-568b-8f74-e8ff0825697b",
-							"name":   "Shahzad",
-							"age":    int64(28),
-						},
-					},
-				},
-			},
-
-			testUtils.UpdateDoc{
-				CollectionID: 0,
-
-				Identity: immutable.Some(2), // Manager still can't update
-
-				DocID: 0,
-
-				Doc: `
-					{
-						"name": "Shahzad Lone"
-					}
-				`,
-
-				ExpectedError: "document not found or not authorized to access",
-			},
-
-			testUtils.DeleteDoc{
-				CollectionID: 0,
-
-				Identity: immutable.Some(2), // Manager still can't delete
-
-				DocID: 0,
-
-				ExpectedError: "document not found or not authorized to access",
-			},
-		},
-	}
-
-	testUtils.ExecuteTestCase(t, test)
-}
-
-func TestACP_OwnerMakesAManagerThatGivesItSelfReadAndWriteAccess_ManagerCanReadAndWrite(t *testing.T) {
-	expectedPolicyID := "fc56b7509c20ac8ce682b3b9b4fdaad868a9c70dda6ec16720298be64f16e9a4"
-
-	test := testUtils.TestCase{
-
-		Description: "Test acp, owner makes a manager that gives itself read and write access",
-
-		SupportedMutationTypes: immutable.Some([]testUtils.MutationType{
-			testUtils.CollectionNamedMutationType,
-			testUtils.CollectionSaveMutationType,
-		}),
-
-		Actions: []any{
-			testUtils.AddPolicy{
-
-				Identity: immutable.Some(1),
-
-				Policy: `
-                    name: Test Policy
-
-                    description: A Policy
-
-                    actor:
-                      name: actor
-
-                    resources:
-                      users:
-                        permissions:
-                          read:
-                            expr: owner + reader + writer
-
-                          write:
-                            expr: owner + writer
-
-                          nothing:
-                            expr: dummy
-
-                        relations:
-                          owner:
-                            types:
-                              - actor
-
-                          reader:
-                            types:
-                              - actor
-
-                          writer:
-                            types:
-                              - actor
-
-                          admin:
-                            manages:
-                              - reader
-                              - writer
-                            types:
-                              - actor
-
-                          dummy:
-                            types:
-                              - actor
-                `,
-
-				ExpectedPolicyID: expectedPolicyID,
-			},
-
-			testUtils.SchemaUpdate{
-				Schema: fmt.Sprintf(`
-						type Users @policy(
-							id: "%s",
-							resource: "users"
-						) {
-							name: String
-							age: Int
-						}
-					`,
-					expectedPolicyID,
-				),
-			},
-
-			testUtils.CreateDoc{
-				Identity: immutable.Some(1),
-
-				CollectionID: 0,
-
-				Doc: `
-					{
-						"name": "Shahzad",
-						"age": 28
-					}
-				`,
-			},
-
-			testUtils.Request{
-				Identity: immutable.Some(2), // This identity (to be manager) can not read yet.
-
-				Request: `
-					query {
-						Users {
-							_docID
-							name
-							age
-						}
-					}
-				`,
-
-				Results: map[string]any{
-					"Users": []map[string]any{}, // Can't see the documents yet
-				},
-			},
-
-			testUtils.UpdateDoc{
-				CollectionID: 0,
-
-				Identity: immutable.Some(2), // Manager can't update yet.
-
-				DocID: 0,
-
-				Doc: `
-					{
-						"name": "Shahzad Lone"
-					}
-				`,
-
-				ExpectedError: "document not found or not authorized to access",
-			},
-
-			testUtils.DeleteDoc{
-				CollectionID: 0,
-
-				Identity: immutable.Some(2), // Manager can't delete yet.
-
-				DocID: 0,
-
-				ExpectedError: "document not found or not authorized to access",
-			},
-
-			testUtils.AddDocActorRelationship{ // Make admin / manager
-				RequestorIdentity: 1,
-
-				TargetIdentity: 2,
-
-				CollectionID: 0,
-
-				DocID: 0,
-
-				Relation: "admin",
-
-				ExpectedExistence: false,
-			},
-
-			testUtils.AddDocActorRelationship{ // Manager makes itself a writer
-				RequestorIdentity: 2,
-
-				TargetIdentity: 2,
-
-				CollectionID: 0,
-
-				DocID: 0,
-
-				Relation: "writer",
-
-				ExpectedExistence: false,
-			},
-
-			// Note: It is not neccesary to make itself a reader, as becoming a writer allows reading.
-			testUtils.AddDocActorRelationship{ // Manager makes itself a reader
-				RequestorIdentity: 2,
-
-				TargetIdentity: 2,
-
-				CollectionID: 0,
-
-				DocID: 0,
-
-				Relation: "reader",
-
-				ExpectedExistence: false,
-			},
-
-			testUtils.UpdateDoc{
-				CollectionID: 0,
-
-				Identity: immutable.Some(2), // Manager can now update.
-
-				DocID: 0,
-
-				Doc: `
-					{
-						"name": "Shahzad Lone"
-					}
-				`,
-			},
-
-			testUtils.Request{
-				Identity: immutable.Some(2), // Manager can read now
-
-				Request: `
-					query {
-						Users {
-							_docID
-							name
-							age
-						}
-					}
-				`,
-
-				Results: map[string]any{
-					"Users": []map[string]any{
-						{
-							"_docID": "bae-9d443d0c-52f6-568b-8f74-e8ff0825697b",
-							"name":   "Shahzad Lone",
-							"age":    int64(28),
-						},
-					},
-				},
-			},
-
-			testUtils.DeleteDoc{
-				CollectionID: 0,
-
-				Identity: immutable.Some(2), // Manager can now delete.
-
-				DocID: 0,
-			},
-
-			testUtils.Request{
-				Identity: immutable.Some(2), // Make sure manager was able to delete the document.
-
-				Request: `
-					query {
-						Users {
-							_docID
-							name
-							age
-						}
-					}
-				`,
-
-				Results: map[string]any{
-					"Users": []map[string]any{},
-				},
-			},
-		},
-	}
-
-	testUtils.ExecuteTestCase(t, test)
-}
-
-func TestACP_ManagerAddsRelationshipWithRelationItDoesNotManageAccordingToPolicy_Error(t *testing.T) {
-	expectedPolicyID := "fc56b7509c20ac8ce682b3b9b4fdaad868a9c70dda6ec16720298be64f16e9a4"
-
-	test := testUtils.TestCase{
-
-		Description: "Test acp, manager adds relationship with relation it does not manage according to policy, error",
-
-		SupportedMutationTypes: immutable.Some([]testUtils.MutationType{
-			testUtils.CollectionNamedMutationType,
-			testUtils.CollectionSaveMutationType,
-		}),
-
-		Actions: []any{
-			testUtils.AddPolicy{
-
-				Identity: immutable.Some(1),
-
-				Policy: `
-                    name: Test Policy
-
-                    description: A Policy
-
-                    actor:
-                      name: actor
-
-                    resources:
-                      users:
-                        permissions:
-                          read:
-                            expr: owner + reader + writer
-
-                          write:
-                            expr: owner + writer
-
-                          nothing:
-                            expr: dummy
-
-                        relations:
-                          owner:
-                            types:
-                              - actor
-
-                          reader:
-                            types:
-                              - actor
-
-                          writer:
-                            types:
-                              - actor
-
-                          admin:
-                            manages:
-                              - reader
-                            types:
-                              - actor
-
-                          dummy:
-                            types:
-                              - actor
-                `,
-
-				ExpectedPolicyID: expectedPolicyID,
-			},
-
-			testUtils.SchemaUpdate{
-				Schema: fmt.Sprintf(`
-						type Users @policy(
-							id: "%s",
-							resource: "users"
-						) {
-							name: String
-							age: Int
-						}
-					`,
-					expectedPolicyID,
-				),
-			},
-
-			testUtils.CreateDoc{
-				Identity: immutable.Some(1),
-
-				CollectionID: 0,
-
-				Doc: `
-					{
-						"name": "Shahzad",
-						"age": 28
-					}
-				`,
-			},
-
-			testUtils.AddDocActorRelationship{ // Make admin / manager
-				RequestorIdentity: 1,
-
-				TargetIdentity: 2,
-
-				CollectionID: 0,
-
-				DocID: 0,
-
-				Relation: "admin",
-
-				ExpectedExistence: false,
-			},
-
-			testUtils.AddDocActorRelationship{ // Admin tries to make another actor a writer
-				RequestorIdentity: 2,
-
-				TargetIdentity: 3,
-
-				CollectionID: 0,
-
-				DocID: 0,
-
-				Relation: "writer",
-
-				ExpectedError: "acp protocol violation",
-			},
-
-			testUtils.Request{
-				Identity: immutable.Some(3), // The other actor can't read
-
-				Request: `
-					query {
-						Users {
-							_docID
-							name
-							age
-						}
-					}
-				`,
-
-				Results: map[string]any{
-					"Users": []map[string]any{},
-				},
-			},
-
-			testUtils.UpdateDoc{
-				CollectionID: 0,
-
-				Identity: immutable.Some(3), // The other actor can not update
-
-				DocID: 0,
-
-				Doc: `
-					{
-						"name": "Shahzad Lone"
-					}
-				`,
-
-				ExpectedError: "document not found or not authorized to access",
-			},
-
-			testUtils.DeleteDoc{
-				CollectionID: 0,
-
-				Identity: immutable.Some(3), // The other actor can not delete
-
-				DocID: 0,
-
-				ExpectedError: "document not found or not authorized to access",
-			},
-		},
-	}
-
-	testUtils.ExecuteTestCase(t, test)
-}
-
-func TestACP_OwnerMakesManagerButManagerCanNotPerformOperations_ManagerCantReadOrWrite(t *testing.T) {
-	expectedPolicyID := "fc56b7509c20ac8ce682b3b9b4fdaad868a9c70dda6ec16720298be64f16e9a4"
-
-	test := testUtils.TestCase{
-
-		Description: "Test acp, owner makes a manager, manager can't read or write",
-
-		SupportedMutationTypes: immutable.Some([]testUtils.MutationType{
-			testUtils.CollectionNamedMutationType,
-			testUtils.CollectionSaveMutationType,
-		}),
-
-		Actions: []any{
-			testUtils.AddPolicy{
-
-				Identity: immutable.Some(1),
-
-				Policy: `
-                    name: Test Policy
-
-                    description: A Policy
-
-                    actor:
-                      name: actor
-
-                    resources:
-                      users:
-                        permissions:
-                          read:
-                            expr: owner + reader + writer
-
-                          write:
-                            expr: owner + writer
-
-                          nothing:
-                            expr: dummy
-
-                        relations:
-                          owner:
-                            types:
-                              - actor
-
-                          reader:
-                            types:
-                              - actor
-
-                          writer:
-                            types:
-                              - actor
-
-                          admin:
-                            manages:
-                              - reader
-                            types:
-                              - actor
-
-                          dummy:
-                            types:
-                              - actor
-                `,
-
-				ExpectedPolicyID: expectedPolicyID,
-			},
-
-			testUtils.SchemaUpdate{
-				Schema: fmt.Sprintf(`
-						type Users @policy(
-							id: "%s",
-							resource: "users"
-						) {
-							name: String
-							age: Int
-						}
-					`,
-					expectedPolicyID,
-				),
-			},
-
-			testUtils.CreateDoc{
-				Identity: immutable.Some(1),
-
-				CollectionID: 0,
-
-				Doc: `
-					{
-						"name": "Shahzad",
-						"age": 28
-					}
-				`,
-			},
-
-			testUtils.AddDocActorRelationship{ // Make admin / manager
-				RequestorIdentity: 1,
-
-				TargetIdentity: 2,
-
-				CollectionID: 0,
-
-				DocID: 0,
-
-				Relation: "admin",
-
-				ExpectedExistence: false,
-			},
-
-			testUtils.Request{
-				Identity: immutable.Some(2), // Manager can not read
-
-				Request: `
-					query {
-						Users {
-							_docID
-							name
-							age
-						}
-					}
-				`,
-
-				Results: map[string]any{
-					"Users": []map[string]any{},
-				},
-			},
-
-			testUtils.UpdateDoc{
-				CollectionID: 0,
-
-				Identity: immutable.Some(2), // Manager can not update.
-
-				DocID: 0,
-
-				Doc: `
-					{
-						"name": "Shahzad Lone"
-					}
-				`,
-
-				ExpectedError: "document not found or not authorized to access",
-			},
-
-			testUtils.DeleteDoc{
-				CollectionID: 0,
-
-				Identity: immutable.Some(2), // Manager can not delete.
-
-				DocID: 0,
-
-				ExpectedError: "document not found or not authorized to access",
-			},
-
-			testUtils.AddDocActorRelationship{ // Manager can manage only.
-				RequestorIdentity: 2,
-
-				TargetIdentity: 3,
-
-				CollectionID: 0,
-
-				DocID: 0,
-
-				Relation: "reader",
-
-				ExpectedExistence: false,
-			},
-		},
-	}
-
-	testUtils.ExecuteTestCase(t, test)
-}
-
-func TestACP_CantMakeRelationshipIfNotOwnerOrManager_Error(t *testing.T) {
-	expectedPolicyID := "fc56b7509c20ac8ce682b3b9b4fdaad868a9c70dda6ec16720298be64f16e9a4"
-
-	test := testUtils.TestCase{
-
-		Description: "Test acp, cant make relation if identity doesn't own or manage object, return error",
+		Description: "Test acp, owner gives read access to another actor, both can read",
 
 		Actions: []any{
 			testUtils.AddPolicy{
@@ -1265,9 +392,9 @@ func TestACP_CantMakeRelationshipIfNotOwnerOrManager_Error(t *testing.T) {
 			},
 
 			testUtils.AddDocActorRelationship{
-				RequestorIdentity: 2, // This identity can not manage as not an admin yet
+				RequestorIdentity: 1,
 
-				TargetIdentity: 3,
+				TargetIdentity: 2,
 
 				CollectionID: 0,
 
@@ -1276,8 +403,405 @@ func TestACP_CantMakeRelationshipIfNotOwnerOrManager_Error(t *testing.T) {
 				Relation: "reader",
 
 				ExpectedExistence: false,
+			},
 
-				ExpectedError: "failed to add document actor relationship with acp",
+			testUtils.Request{
+				Identity: immutable.Some(2), // Now this identity can read.
+
+				Request: `
+					query {
+						Users {
+							_docID
+							name
+							age
+						}
+					}
+				`,
+
+				Results: map[string]any{
+					"Users": []map[string]any{
+						{
+							"_docID": "bae-9d443d0c-52f6-568b-8f74-e8ff0825697b",
+							"name":   "Shahzad",
+							"age":    int64(28),
+						},
+					},
+				},
+			},
+
+			testUtils.Request{
+				Identity: immutable.Some(1), // And so can the owner (ownership not transferred).
+
+				Request: `
+					query {
+						Users {
+							_docID
+							name
+							age
+						}
+					}
+				`,
+
+				Results: map[string]any{
+					"Users": []map[string]any{
+						{
+							"_docID": "bae-9d443d0c-52f6-568b-8f74-e8ff0825697b",
+							"name":   "Shahzad",
+							"age":    int64(28),
+						},
+					},
+				},
+			},
+		},
+	}
+
+	testUtils.ExecuteTestCase(t, test)
+}
+
+func TestACP_OwnerGivesOnlyReadAccessToAnotherActor_OtherActorCanReadButNotUpdate(t *testing.T) {
+	expectedPolicyID := "fc56b7509c20ac8ce682b3b9b4fdaad868a9c70dda6ec16720298be64f16e9a4"
+
+	test := testUtils.TestCase{
+
+		Description: "Test acp, owner gives read access to another actor, but the other actor can't update",
+
+		SupportedMutationTypes: immutable.Some([]testUtils.MutationType{
+			testUtils.CollectionNamedMutationType,
+			testUtils.CollectionSaveMutationType,
+		}),
+
+		Actions: []any{
+			testUtils.AddPolicy{
+
+				Identity: immutable.Some(1),
+
+				Policy: `
+                    name: Test Policy
+
+                    description: A Policy
+
+                    actor:
+                      name: actor
+
+                    resources:
+                      users:
+                        permissions:
+                          read:
+                            expr: owner + reader + writer
+
+                          write:
+                            expr: owner + writer
+
+                          nothing:
+                            expr: dummy
+
+                        relations:
+                          owner:
+                            types:
+                              - actor
+
+                          reader:
+                            types:
+                              - actor
+
+                          writer:
+                            types:
+                              - actor
+
+                          admin:
+                            manages:
+                              - reader
+                            types:
+                              - actor
+
+                          dummy:
+                            types:
+                              - actor
+                `,
+
+				ExpectedPolicyID: expectedPolicyID,
+			},
+
+			testUtils.SchemaUpdate{
+				Schema: fmt.Sprintf(`
+						type Users @policy(
+							id: "%s",
+							resource: "users"
+						) {
+							name: String
+							age: Int
+						}
+					`,
+					expectedPolicyID,
+				),
+			},
+
+			testUtils.CreateDoc{
+				Identity: immutable.Some(1),
+
+				CollectionID: 0,
+
+				Doc: `
+					{
+						"name": "Shahzad",
+						"age": 28
+					}
+				`,
+			},
+
+			testUtils.Request{
+				Identity: immutable.Some(2), // This identity can not read yet.
+
+				Request: `
+					query {
+						Users {
+							_docID
+							name
+							age
+						}
+					}
+				`,
+
+				Results: map[string]any{
+					"Users": []map[string]any{}, // Can't see the documents yet
+				},
+			},
+
+			testUtils.UpdateDoc{ // Since it can't read, it can't update either.
+				CollectionID: 0,
+
+				Identity: immutable.Some(2),
+
+				DocID: 0,
+
+				Doc: `
+					{
+						"name": "Shahzad Lone"
+					}
+				`,
+
+				ExpectedError: "document not found or not authorized to access",
+			},
+
+			testUtils.AddDocActorRelationship{
+				RequestorIdentity: 1,
+
+				TargetIdentity: 2,
+
+				CollectionID: 0,
+
+				DocID: 0,
+
+				Relation: "reader",
+
+				ExpectedExistence: false,
+			},
+
+			testUtils.Request{
+				Identity: immutable.Some(2), // Now this identity can read.
+
+				Request: `
+					query {
+						Users {
+							_docID
+							name
+							age
+						}
+					}
+				`,
+
+				Results: map[string]any{
+					"Users": []map[string]any{
+						{
+							"_docID": "bae-9d443d0c-52f6-568b-8f74-e8ff0825697b",
+							"name":   "Shahzad",
+							"age":    int64(28),
+						},
+					},
+				},
+			},
+
+			testUtils.UpdateDoc{ // But this actor still can't update.
+				CollectionID: 0,
+
+				Identity: immutable.Some(2),
+
+				DocID: 0,
+
+				Doc: `
+					{
+						"name": "Shahzad Lone"
+					}
+				`,
+
+				ExpectedError: "document not found or not authorized to access",
+			},
+		},
+	}
+
+	testUtils.ExecuteTestCase(t, test)
+}
+
+func TestACP_OwnerGivesOnlyReadAccessToAnotherActor_OtherActorCanReadButNotDelete(t *testing.T) {
+	expectedPolicyID := "fc56b7509c20ac8ce682b3b9b4fdaad868a9c70dda6ec16720298be64f16e9a4"
+
+	test := testUtils.TestCase{
+
+		Description: "Test acp, owner gives read access to another actor, but the other actor can't delete",
+
+		Actions: []any{
+			testUtils.AddPolicy{
+
+				Identity: immutable.Some(1),
+
+				Policy: `
+                    name: Test Policy
+
+                    description: A Policy
+
+                    actor:
+                      name: actor
+
+                    resources:
+                      users:
+                        permissions:
+                          read:
+                            expr: owner + reader + writer
+
+                          write:
+                            expr: owner + writer
+
+                          nothing:
+                            expr: dummy
+
+                        relations:
+                          owner:
+                            types:
+                              - actor
+
+                          reader:
+                            types:
+                              - actor
+
+                          writer:
+                            types:
+                              - actor
+
+                          admin:
+                            manages:
+                              - reader
+                            types:
+                              - actor
+
+                          dummy:
+                            types:
+                              - actor
+                `,
+
+				ExpectedPolicyID: expectedPolicyID,
+			},
+
+			testUtils.SchemaUpdate{
+				Schema: fmt.Sprintf(`
+						type Users @policy(
+							id: "%s",
+							resource: "users"
+						) {
+							name: String
+							age: Int
+						}
+					`,
+					expectedPolicyID,
+				),
+			},
+
+			testUtils.CreateDoc{
+				Identity: immutable.Some(1),
+
+				CollectionID: 0,
+
+				Doc: `
+					{
+						"name": "Shahzad",
+						"age": 28
+					}
+				`,
+			},
+
+			testUtils.Request{
+				Identity: immutable.Some(2), // This identity can not read yet.
+
+				Request: `
+					query {
+						Users {
+							_docID
+							name
+							age
+						}
+					}
+				`,
+
+				Results: map[string]any{
+					"Users": []map[string]any{}, // Can't see the documents yet
+				},
+			},
+
+			testUtils.DeleteDoc{ // Since it can't read, it can't delete either.
+				CollectionID: 0,
+
+				Identity: immutable.Some(2),
+
+				DocID: 0,
+
+				ExpectedError: "document not found or not authorized to access",
+			},
+
+			testUtils.AddDocActorRelationship{
+				RequestorIdentity: 1,
+
+				TargetIdentity: 2,
+
+				CollectionID: 0,
+
+				DocID: 0,
+
+				Relation: "reader",
+
+				ExpectedExistence: false,
+			},
+
+			testUtils.Request{
+				Identity: immutable.Some(2), // Now this identity can read.
+
+				Request: `
+					query {
+						Users {
+							_docID
+							name
+							age
+						}
+					}
+				`,
+
+				Results: map[string]any{
+					"Users": []map[string]any{
+						{
+							"_docID": "bae-9d443d0c-52f6-568b-8f74-e8ff0825697b",
+							"name":   "Shahzad",
+							"age":    int64(28),
+						},
+					},
+				},
+			},
+
+			testUtils.DeleteDoc{ // But this actor still can't delete.
+				CollectionID: 0,
+
+				Identity: immutable.Some(2),
+
+				DocID: 0,
+
+				ExpectedError: "document not found or not authorized to access",
 			},
 		},
 	}
